// src/app/core/model/about/about.ts
import { Address } from '../address/address';

export class About {
  id: any;
  content: any;
  description: any =
    'Công ty Cổ phần EcoIT được đăng ký thành lập ngày 29/4/2010 là một Công ty tin học chuyên cung cấp phần mềm, dịch vụ, hệ thống và các giải pháp về Công nghệ thông tin được sáng lập và lãnh đạo bởi các chuyên gia quản lý, chuyên gia công nghệ có kinh nghiệm lâu năm trong lĩnh vực Công nghệ thông tin có cùng ý tưởng đến từ nhiều tổ chức, doanh nghiệp trong các lĩnh vực khác nhau.';
  videoLINK: any = 'https://www.youtube.com/watch?v=cG-zYVv4fVw';
  phone: any = ['(+84) 935 772 929'];
  fax: any = ['(024) 62750064'];
  email: any = ['contact@ecoit.com.vn'];
  address: any;
}